# Desktop/airbnb-project/src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Rest Clon de Airbnb",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/users": {
            "get": {
                "description": "Get All Users",
                "tags": ["Users", "v1"],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
...
    "/api/v1/accommodations": {
        "get": {
            "description": "Get all accommodations",
            "tags": ["Accommodations", "v1"],
            "responses": {
                "200": {
                    "description": "successful operation"
                },
                "400": {
                    "description": "Invalid value"
                }
            }
        },
        "post":{
            "parameters":[
                {
                    "name": "id",
                    "in": "path",
                    "description": "accommodation ID",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "body",
                    "in": "body",
                    "description": "The data object for accommodation",
                    "required": true,
                    "type": "string"
                }
            ],
            "description": "Create accommodation",
            "tags": ["Accommodations", "v1"],
            "responses": {
                "201": {
                    "description": "successful operation"
                },
                "400": {
                    "description": "Invalid value"
                }
            }
        }
    },
...
}
}